refactor: 코드 리뷰 때 지적 받은 부분 재구현
상수로 지정해주고, view와 domain 패키지를 관리하는 Lotto 클래스를 최상단 폴더에 위치시켰다.
