+++ src/main/java/kr/codesquad/domain/LottoGenerator.java
@@ -1,6 +1,5 @@
 package kr.codesquad.domain;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
